requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.txt
#
adal==1.2.1               # via -r requirements.txt, botframework-connector
aiohttp==3.6.2            # via -r requirements.txt
aiounittest==1.3.0        # via -r requirements.txt, botbuilder-dialogs
asgiref==3.2.10           # via -r requirements.txt
astroid==2.4.2            # via -r requirements.txt, pylint
async-timeout==3.0.1      # via -r requirements.txt, aiohttp
attrs==19.3.0             # via -r requirements.txt, aiohttp
babel==2.7.0              # via -r requirements.txt, botbuilder-dialogs
botbuilder-core==4.9.1    # via -r requirements.txt, botbuilder-dialogs
botbuilder-dialogs==4.9.1  # via -r requirements.txt
botbuilder-schema==4.9.1  # via -r requirements.txt, botbuilder-core, botbuilder-dialogs, botframework-connector
botframework-connector==4.9.1  # via -r requirements.txt, botbuilder-core, botbuilder-dialogs
certifi==2020.6.20        # via -r requirements.txt, msrest, requests
cffi==1.14.0              # via -r requirements.txt, cryptography
chardet==3.0.4            # via -r requirements.txt, aiohttp, requests
cryptography==2.8         # via -r requirements.txt, adal, botframework-connector, pyjwt
datedelta==1.3            # via -r requirements.txt, recognizers-text-date-time
django-dbbackup==3.3.0    # via -r requirements.txt
django==2.1.2             # via -r requirements.txt, django-dbbackup
emoji==0.5.4              # via -r requirements.txt, recognizers-text
grapheme==0.6.0           # via -r requirements.txt, recognizers-text-choice
idna==2.8                 # via -r requirements.txt, requests, yarl
isodate==0.6.0            # via -r requirements.txt, msrest
isort==4.3.21             # via -r requirements.txt, pylint
jsonpickle==1.2           # via -r requirements.txt, botbuilder-core
lazy-object-proxy==1.4.3  # via -r requirements.txt, astroid
mccabe==0.6.1             # via -r requirements.txt, pylint
msal==1.2.0               # via -r requirements.txt, botframework-connector
msrest==0.6.10            # via -r requirements.txt, botbuilder-schema, botframework-connector
multidict==4.7.6          # via -r requirements.txt, aiohttp, yarl
multipledispatch==0.6.0   # via -r requirements.txt, recognizers-text
oauthlib==3.1.0           # via -r requirements.txt, requests-oauthlib
psycopg2-binary==2.8.5    # via -r requirements.txt
pycparser==2.20           # via -r requirements.txt, cffi
pyjwt[crypto]==1.5.3      # via -r requirements.txt, adal, botframework-connector, msal
pylint==2.5.3             # via -r requirements.txt
python-dateutil==2.8.1    # via -r requirements.txt, adal
python-dotenv==0.14.0     # via -r requirements.txt
pytz==2020.1              # via -r requirements.txt, babel, django, django-dbbackup
recognizers-text-choice==1.0.2a2  # via -r requirements.txt, botbuilder-dialogs
recognizers-text-date-time==1.0.2a2  # via -r requirements.txt, botbuilder-dialogs
recognizers-text-number-with-unit==1.0.2a2  # via -r requirements.txt, botbuilder-dialogs, recognizers-text-date-time
recognizers-text-number==1.0.2a2  # via -r requirements.txt, botbuilder-dialogs, recognizers-text-date-time, recognizers-text-number-with-unit
recognizers-text==1.0.2a2  # via -r requirements.txt, botbuilder-dialogs, recognizers-text-choice, recognizers-text-date-time, recognizers-text-number, recognizers-text-number-with-unit
regex==2019.8.19          # via botbuilder-dialogs, recognizers-text-choice, recognizers-text-date-time, recognizers-text-number, recognizers-text-number-with-unit
requests-oauthlib==1.3.0  # via -r requirements.txt, msrest
requests==2.22.0          # via -r requirements.txt, adal, botframework-connector, msal, msrest, requests-oauthlib
six==1.15.0               # via -r requirements.txt, astroid, cryptography, django-dbbackup, isodate, multipledispatch, python-dateutil, websocket-client
sqlparse==0.3.1           # via -r requirements.txt
toml==0.10.1              # via -r requirements.txt, pylint
urllib3==1.25.9           # via -r requirements.txt, requests
websocket-client==0.57.0  # via -r requirements.txt
whitenoise==5.1.0         # via -r requirements.txt
wrapt==1.12.1             # via -r requirements.txt, astroid
yarl==1.4.2               # via -r requirements.txt, aiohttp
